Added game and moved implementations

+++ src/heuristic.rs
@@ -2,7 +2,7 @@ use crate::{player::Player, state::State};
 
 
 impl State {
-    pub fn heuristic(&self, _player: Player) -> f64 {
+    pub fn heuristic(&self, _player: Player) -> f32 {
         0.0
     }    
 }
